requirements.txt
asgiref==3.6.0
decorator==5.1.1
Django==4.1.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
validators==0.20.0
